pyproject.toml
[tool.poetry]
name = "kafka-connect-watcher"
version = "0.1.0"
description = ""
authors = ["John \"Preston\" Mille <john@ews-network.net>"]
readme = "README.rst"
packages = [{ include = "kafka_connect_watcher" }]
include = [
  "kafka_connect_watcher/**/*.json",
]
exclude = ["*.pyc", "*~", "*pycache*"]

[tool.poetry.scripts]
kafka_connect_watcher = "kafka_connect_watcher.cli:start_watcher"
kafka-connect-watcher = "kafka_connect_watcher.cli:start_watcher"

[tool.poetry.dependencies]
python = "^3.8"
compose-x-common = "^1.2.6"
kafka-connect-api = "^0.5.0"
pyyaml = "^6.0"
jsonschema = "^4.17.3"
importlib-resources = "^5.10.2"
prometheus-client = "^0.15.0"
aws-embedded-metrics = "^3.0.0"


[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
isort = "^5.11.4"
pre-commit = "^2.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
known_first_party = "kelvin"

[tool.coverage.report]
exclude_lines = [
  "if __name__ == '__main__'"
]

[tool.coverage.run]
omit = [
  "*/cli.py"
]
